chore: use autoload and remove dead code
