+++ lib/metanorma/plugin/plantuml/config.rb
@@ -26,18 +26,6 @@ module Metanorma
           options
         end
       end
-
-      class << self
-        attr_writer :configuration
-      end
-
-      def self.configuration
-        @configuration ||= Config.new
-      end
-
-      def self.configure
-        yield(configuration)
-      end
     end
   end
 end
